Fix: Correct theme definition, imports, and types in MainActivity

+++ app/src/main/java/com/subho/aurabeat/MainActivity.kt
@@ -59,7 +59,7 @@ class MainActivity : ComponentActivity() {
                     onPlayPauseClick = { musicPlayer.togglePlayPause() },
                     onNextClick = { musicPlayer.playNext() },
                     onPreviousClick = { musicPlayer.playPrevious() },
-                    onSeek = { progress -> musicPlayer.seekTo(progress) }
+                    onSeek = { progress -> musicPlayer.seekTo(progress.toInt()) }
                 )
             }
         }
